requirements.txt
langchain
langgraph
langchain-core
langchain-openai
langchain-text-splitters
openai
streamlit
python-dotenv
pydantic
pytest
email-validator
